# docs/swagger.yaml
basePath: /v1
definitions:
  entity.AutoMessageSender:
    properties:
      action:
        example: start/stop
        type: string
    type: object
  entity.Message:
    properties:
      content:
        type: string
      recipient_phone:
        type: string
    type: object
  v1.autoMessageResponse:
    properties:
      Description:
        type: string
      status:
        type: string
    type: object
  v1.messageResponse:
    properties:
      Messages:
        items:
          $ref: '#/definitions/entity.Message'
        type: array
      count:
        description: Number of messages in the response
        type: integer
      error:
        description: Error message, optional field
        type: string
      status:
        description: Status of the response (e.g., "success", "empty")
        type: string
    type: object
  v1.response:
    properties:
      error:
        example: message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: The system sends 2 messages every 2 minutes.
  title: Automatic Message Sender API
  version: "1.0"
paths:
  /automatic-message-sender:
    post:
      consumes:
      - application/json
      description: Starts or stops the automatic message sender scheduler based on
        the provided action ("start" or "stop").
      operationId: control-message-scheduler
      parameters:
      - description: Action for the scheduler (start or stop)
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/entity.AutoMessageSender'
      produces:
      - application/json
      responses:
        "200":
          description: Operation completed successfully with a success message
          schema:
            $ref: '#/definitions/v1.autoMessageResponse'
        "400":
          description: Invalid input or unsupported action
          schema:
            $ref: '#/definitions/v1.autoMessageResponse'
        "500":
          description: Internal server error while starting or stopping the scheduler
          schema:
            $ref: '#/definitions/v1.autoMessageResponse'
      summary: Control the automatic message sender scheduler
      tags:
      - Scheduler
  /messages:
    post:
      consumes:
      - application/json
      description: Add a new message to the database
      operationId: insert-message
      parameters:
      - description: Message to insert
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/entity.Message'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Insert a message
      tags:
      - messages
  /messages/sent:
    get:
      consumes:
      - application/json
      description: Retrieve all messages with status "sent" from the database
      operationId: get-sent-messages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.messageResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Retrieve sent messages
      tags:
      - messages
swagger: "2.0"